feat: refactored MessagesContext
