+++ packages/app/public/config.template.js
@@ -5,6 +5,7 @@ window.OIDC_REDIRECT_URI = "${OIDC_REDIRECT_URI}";
 window.GRAPHQL_URI = "${GRAPHQL_URI}";
 window.REST_URI = "${REST_URI}";
 window.SHOW_INHABITANT_AMOUNT = "${SHOW_INHABITANT_AMOUNT}";
+window.MESSAGE_COUNT_ENABLE = "${MESSAGE_COUNT_ENABLE}" === "true";
 window.MESSAGE_COUNT_POLLING_INTERVAL = Number(
   "${MESSAGE_COUNT_POLLING_INTERVAL}",
 );
