+++ packages/app/src/App.tsx
@@ -6,9 +6,8 @@ import { LocalizationProvider } from "@nl-portal/nl-portal-localization";
 import { ApiProvider } from "@nl-portal/nl-portal-api";
 import {
   Layout,
-  MessagesProvider,
+  AppProvider,
   NotificationProvider,
-  useEnableMessagesCount,
 } from "@nl-portal/nl-portal-user-interface";
 import { CUSTOM_MESSAGES } from "./i18n/custom-messages/custom-messages";
 import HeaderLogo from "./assets/header-logo.svg";
@@ -27,8 +26,6 @@ const authenticationMethods = {
 };
 
 const App = () => {
-  const enableMessagesCount = useEnableMessagesCount(menuItems);
-
   return (
     <div className={config.THEME_CLASS}>
       <LocalizationProvider customMessages={CUSTOM_MESSAGES}>
@@ -45,7 +42,7 @@ const App = () => {
             restUri={config.REST_URI}
           >
             <NotificationProvider>
-              <MessagesProvider enableMessagesCount={enableMessagesCount}>
+              <AppProvider>
                 <Layout
                   navigationItems={menuItems}
                   paths={paths}
@@ -56,7 +53,7 @@ const App = () => {
                   facet={<img src={Facet} alt="facet" />}
                   footer={footerData}
                 />
-              </MessagesProvider>
+              </AppProvider>
             </NotificationProvider>
           </ApiProvider>
         </OidcProvider>
