+++ packages/app/src/constants/config.ts
@@ -18,6 +18,7 @@ export const config = {
   OVERVIEW_CURRENT_CASES_PREVIEW_LENGTH:
     window.OVERVIEW_CURRENT_CASES_PREVIEW_LENGTH,
   THEME_CLASS: window.THEME_CLASS,
+  MESSAGE_COUNT_ENABLE: window.MESSAGE_COUNT_ENABLE,
   MESSAGE_COUNT_POLLING_INTERVAL: window.MESSAGE_COUNT_POLLING_INTERVAL,
   CASES_PARTIAL_SEARCH: window.CASES_PARTIAL_SEARCH,
 };
