+++ packages/authentication/src/window.d.ts
@@ -16,6 +16,7 @@ export declare global {
     OVERVIEW_SHOW_CURRENT_CASES_PREVIEW: string;
     OVERVIEW_CURRENT_CASES_PREVIEW_LENGTH: string;
     THEME_CLASS: string;
+    MESSAGE_COUNT_ENABLE: string;
     MESSAGE_COUNT_POLLING_INTERVAL: number;
     CASES_PARTIAL_SEARCH: boolean;
   }
