+++ packages/user-interface/src/index.ts
@@ -31,7 +31,6 @@ export * from "./interfaces/paths";
 export * from "./interfaces/portal-footer";
 
 export { default as useDownload } from "./hooks/useDownload";
-export { default as useEnableMessagesCount } from "./hooks/useEnableMessagesCount";
 export { default as useOgonePayment } from "./hooks/useOgonePayment";
 export { default as useOgonePaymentRegistration } from "./hooks/useOgonePaymentRegistration";
 export * from "./hooks/useOgonePaymentRegistration";
@@ -54,7 +53,7 @@ export { default as ThemeDetailsPage } from "./pages/ThemeDetailsPage";
 export { default as ThemeOverviewPage } from "./pages/ThemeOverviewPage";
 export { default as ThemeSubPage } from "./pages/ThemeSubPage";
 
-export { default as MessagesContext } from "./contexts/MessagesContext";
-export * from "./contexts/MessagesContext";
+export { default as AppContext } from "./contexts/AppContext";
+export * from "./contexts/AppContext";
 export { default as NotificationContext } from "./contexts/NotificationContext";
 export * from "./contexts/NotificationContext";
